Bänder makes little sense in German.
Bänder = Tapes but the context is not given here. Literally a physical tape. That makes little sense. I guess it's better to use "file" in that context. (If I understood correctly that this is the setting where you can decide to split the audio book every x minutes into a new file)